Backporting fixes from 6.0 trunk to 5.1

+++ as/arm.c
@@ -17839,7 +17839,9 @@ md_apply_fix (fixS *	fixP,
        {
          bfd_vma insn;
          bfd_vma encoded_addend;
-         bfd_vma addend_abs = abs (value);
+          // Patch 12/29/2014 Shutup compile warning
+          // NOTE This value is casted it int32_t by upstream!
+         bfd_vma addend_abs = abs ((int32_t) value);
 
          /* Check that the absolute value of the addend can be
             expressed as an 8-bit constant plus a rotation.  */
@@ -17884,7 +17886,9 @@ md_apply_fix (fixS *	fixP,
 #endif
         {
           bfd_vma insn;
-          bfd_vma addend_abs = abs (value);
+          // Patch 12/29/2014 Shutup compile warning
+          // NOTE This value is casted it int32_t by upstream!
+          bfd_vma addend_abs = abs ((int32_t) value);
 
           /* Check that the absolute value of the addend can be
              encoded in 12 bits.  */
@@ -17927,7 +17931,9 @@ md_apply_fix (fixS *	fixP,
 #endif
         {
           bfd_vma insn;
-          bfd_vma addend_abs = abs (value);
+          // Patch 12/29/2014 Shutup compile warning
+          // NOTE This value is casted it int32_t by upstream!
+          bfd_vma addend_abs = abs ((int32_t) value);
 
           /* Check that the absolute value of the addend can be
              encoded in 8 bits.  */
@@ -17971,7 +17977,9 @@ md_apply_fix (fixS *	fixP,
 #endif
         {
           bfd_vma insn;
-          bfd_vma addend_abs = abs (value);
+          // Patch 12/29/2014 Shutup compile warning
+          // NOTE This value is casted it int32_t by upstream!
+          bfd_vma addend_abs = abs ((int32_t) value);
 
           /* Check that the absolute value of the addend is a multiple of
              four and, when divided by four, fits in 8 bits.  */
